+++ CHANGELOG.md
@@ -2,7 +2,10 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
- * Bumped runtimeVersion to `2023.02.28`.
+ * Bumped runtimeVersion to `2023.02.29`.
+ * Upgraded stable Dart analysis SDK to `2.19.3`.
+ * Upgraded stable Flutter analysis SDK to `3.7.5`.
+ * Upgraded preview Dart analysis SDK to `3.0.0-218.1.dev`.
  * Removed `required: true` from the `UserSession.email` field.
 
 ## `20230223t105400-all`